rename: UserSearchBar -> UserSearch
